Create project structure.
Starting point for the project. Initial commits will probably only be to the
README.md, to capture notes as I'm finishing up Strata, and considering more
carefully how that b&#x2011;tree will be used.

Creating an MVCC database is a good way to exercise Strata. It makes the point
that Strata is not supposed to be used directly as a general purpose database,
but as an indexed data store.