Implement snapshot `get`.
Created a common subclass for `Mutator` and `Snapshot` called
`Transaction` that can contain the functions common to both. For now it
contains the `async get` that calls the internal synchronous `_get`
which is different for `Mutator` and `Snapshot`.

The `Snapshot._get()` is simple. It merely called `Amalgamator.get()`.
The `Mutator._get()` has to check the in-memory stages for values.

Closes #78.