Remove a temporary re-assignment of `this`.
I refactored. An `InnerIterator` used to reference the `OuterIterator`.
These inner and outer JavaScript iterators are now a thin wrapper around
an iteration machine that implements both inner and outer iteration.
During this refactor where the inner iteration moved from the
`InnerIterator` into an object that once was the `OuterIterator`, I
assigned `this` to `outer` because too much was changing, there was an
`outer` I didn't rename to `this` that I couldn't see,  and I was
getting exceptions I didn't want to trace.