Create common `_iterator` function.
I was able to make this complicated function common to both `Snapshot`
and `Mutator`. Parameterized the iterator constructor so that we
ultimately build either `SnapshotIterator` or `MutatorIterator`. Made
the mutator structure the result of calling a virtual function (defined
in a concrete class) where the `Mutator` creates a mutator structure and
the `Snapshot` creates a bogus mutator structure suitable for read-only.

See #76.