Upgrade `amalgamate` to `0.1.35`.
Amalgamate version 0.1.33 allows for a user specified group function for
`map` which we'll use to implement map against indices which may have
more than value for a given key. It also allows us to map against
partial keys. Also changed the results so that the actual key is
returned and the sought value and extracted key are gathered in a sought
structure.

Amalgamate version 0.1.34 exposes the comparators so we can remove the
under bar and stop sneaking around inside our `Amalgamator`s.

Amalgamate version 0.1.25 fixed a bug in the creation of the
aforementioned sought structure.

Closes #81.