Implement snapshot map.
Upgraded `amalgamate` to `0.1.36` to get new structure returned from
`Amalgamate.map()`. Remove the header information and the result is
roughly the same structure, an entry for each key in the sought set with
an array of zero, one or more items that matched the key. The key can be
a partial key so their can be multiple results for for stores and
indices. The results include the key and value. For indices each entry
will have the full index key, not the primary key of the stored object.

https://github.com/bigeasy/amalgamate/releases/tag/v0.1.36

Closes #78.