Upgrade `amalgamate` to `0.1.37`.
Amalgamate 0.1.37 uses the new iterators that have a type property, one
of `mvcc.FORWARD`, `mvcc.REVERSE` or `mvcc.MAP`. This is used to
determine what sort of iterator to create if a decorator iterator's
behavior is based on direction.

Amalgamate 0.1.37 introduces an async static open function instead of
waiting for a `ready` `Promise` property.

https://github.com/bigeasy/amalgamate/releases/tag/v0.1.37

Closes #84.