Implement snapshot index get.
Implement a get of an index from a snapshot. Created a `constrain`
iterator that can be used as a general purpose utility to find the end
of an iteration. Also created `satiate` to ensure that the array
returned had a minimum number of items. This is useful for `get` to
ensure that you get at least one item and don't have to add a loop to
skip an initial empty array.

Index get first gets the item out of the index, then uses the foreign
key part of the compound key to get the actual item. It returns the
actual item. It doesn't make sense for use with an index if the index is
not unique. There is currently no plan to enforce unique indices.

Closes #79.