Merge primary and indices at once.
We're going to track the heft of a primary and its indices and merge
them all when a boundary is reached. This instead of considering each
primary and index separately. Necessary to keep on top of index updates,
we want to put our deletes in the same batch as our updates for the
indices. Indices require deletes and inserts as opposed to just inserts
because their key is extracted from the primary record and is liable to
change. Updating the primary record on the other hand is always an
insert because the primary record is defined by its key, the key does
not change.

Closes #90.