Update old index entry when updating object.
Using the `map` utility to get a map of existing objects from a
snapshot. Using the map to extract the existing keys. Out concurrency
strategy had better work or the index updates are going to be incorrect.
If the concurrency strategy doesn't the primary object updates are also
incorrect, so we'll see soon enough.

Out index get implementation was broken, extracting the key from the
versioned key instead of from the combined index key, but the tests
where passing because the value extracted just happened to be the
primary key value of the correct primary object.

Closes #68.