Major dependency upgrade.
Went back down the stack of depenencies to add a write-ahead log and to
sort out application shutdown. This commit has all tests on an initial
empty Memento working. Going to revisit reopening with next commit due
to change in how commits are stored. They are now stored in the
write-ahead log.