Create static `slurp` and `first`.
Going to forgo syntax bashing `min`/`max` limit, filter and the like for
now and just add some static functions. Let's build some applications
before we start trying to cram the interface with features. We
over-interfaced Destructible and ultimately ended up walking it all back
down to `durable` and `ephemeral` and that's it.