Replace use actual `Map` for mutator index map.
The IndexedDB Web Platform Tests monkey patch Object to test that
cloning and key setting do not use properties that are not own
properties of an object. This breaks Memento that is using objects as
hash tables. For starters we can replace the store's index map and the
mutator's index map with actual Maps. Further testing might reveal
further weaknesses.

We start by replacing the map used to track the indices for a mutating
store. Next we replace the map that is part of the in-memory schema.

See #138.